pyproject.toml
[tool.poetry]
name = "temba"
version = "7.5.103"
description = "Hosted service for visually building interactive messaging applications"
authors = ["Nyaruka <code@nyaruka.com>"]
license = "AGPL-3"

[tool.poetry.urls]
repository = "http://github.com/rapidpro/rapidpro"

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.0.7"
django-compressor = "^3.1"
django-countries = "^7.0"
django-hamlpy = "^1.4.4"
django-mptt = "^0.12.0"
django-redis = "^4.12.1"
django-storages = "^1.11.1"
django-timezone-field = "^4.1.1"
djangorestframework = "^3.12.2"
dj-database-url = "^0.5.0"
smartmin = "^4.1.0"
celery = "^5.1.0"
redis = "^3.5.3"
elasticsearch = "^7.11.0"
elasticsearch-dsl = "^7.3.0"
boto3 = "^1.17.21"
cryptography = "3.4.7"
vonage = "2.5.2"
pyfcm = "^1.5.1"
pyotp = "2.4.1"
python-telegram-bot = "^13.14"
sentry-sdk = "^1.0.0"
twilio = "6.24.0"
twython = "3.5.0"
geojson = "^2.5.0"
Markdown = "^3.3.4"
polib = "^1.1.0"
pyexcel = "^0.6.6"
pyexcel-xls = "^0.6.2"
pyexcel-xlsx = "^0.6.0"
python-magic = "^0.4.22"
xlsxlite = "^0.2.0"
colorama = "^0.4.4"
gunicorn = "^20.0.4"
iptools = "^0.7.0"
iso-639 = "^0.4.5"
iso8601 = "^0.1.14"
phonenumbers = "*"
pycountry = "^20.7.3"
python-dateutil = "2.2"
psycopg2-binary = "^2.9.1"
packaging = "^20.9"
requests-toolbelt = "^0.9.1"
chardet = "^4.0.0"
openpyxl = "^3.0.9"
ffmpeg-python = "^0.2.0"
slack-sdk = "3.17.0"
djangorestframework-xml = "^2.0.0"
python-json-logger = "^2.0.4"

[tool.poetry.dev-dependencies]
black = "^22.1.0"
coverage = "^5.5"
flake8 = "^4.0.1"
isort = "^5.10.1"
codecov = "^2.1.11"
responses = "^0.12.1"
beautifulsoup4 = "^4.9.3"

[tool.black]
line-length = 119

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
